SPOP-432 Fix some power up configuration issues and change the data to use ints instead of floats for the power ups since that number is supposed to be max fills not a rate.

+++ Assets/Scripts/LevelEditor/Properties/LevelProperties.cs
@@ -14,13 +14,13 @@ namespace LevelEditor.Properties
         [PropertyDisplayAttribute(
             new string[] { "Blue PowerUp Fills", "Yellow PowerUp Fills", "Red PowerUp Fills", "Green PowerUp Fills", "Pink PowerUp Fills", "Purple PowerUp Fills" },
             new string[] { "3484d8", "ea9115", "d22d2b", "11952a", "e746a1", "891fc4" })]
-        public float[] PowerUpFills { get; set; }
+        public int[] PowerUpFills { get; set; }
 
         public LevelProperties()
         {
             ShotCount = 20;
             StarValues = new int[] { 100, 500, 1000 };
-            PowerUpFills = new float[6];
+            PowerUpFills = new int[6];
         }
 
         public void FromLevelData(LevelData data)
